main.py add docstrings
